reintegrate `watcher` crate into `tauri` crate
