+++ crates/gitbutler-tauri/Cargo.toml
@@ -57,6 +57,8 @@ tracing = "0.1.40"
 tracing-appender = "0.2.3"
 tracing-subscriber = "0.3.17"
 gitbutler-core.workspace = true
+gitbutler-watcher.workspace = true
+gitbutler-analytics.workspace = true
 
 [dependencies.tauri]
 version = "1.6.1"
