--- crates/gitbutler-tauri/tests/app.rs
@@ -1,1 +1,0 @@
-mod watcher;
