--- crates/gitbutler-tauri/tests/watcher/mod.rs
@@ -1,1 +1,0 @@
-mod handler;
